# openapi.yaml
components:
  schemas:
...
    Body_list_tickets_api_v1_tickets__get:
      properties:
        category:
          anyOf:
          - items:
              $ref: '#/components/schemas/TicketCategorySchema'
            type: array
          - type: 'null'
          title: Category
        priority:
          anyOf:
          - items:
              $ref: '#/components/schemas/TicketPrioritySchema'
            type: array
          - type: 'null'
          title: Priority
        status:
          anyOf:
          - items:
              $ref: '#/components/schemas/TicketStatusSchema'
            type: array
          - type: 'null'
          title: Status
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        urgency:
          anyOf:
          - items:
              $ref: '#/components/schemas/TicketPrioritySchema'
            type: array
          - type: 'null'
          title: Urgency
      title: Body_list_tickets_api_v1_tickets__get
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PaginatedResponse:
      description: Standard paginated response wrapper
      properties:
        has_next:
          description: Whether there are more pages
          title: Has Next
          type: boolean
        has_prev:
          description: Whether there are previous pages
          title: Has Prev
          type: boolean
        items:
          description: List of items
          items: {}
          title: Items
          type: array
        page:
          description: Current page number
          title: Page
          type: integer
        pages:
          description: Total number of pages
          title: Pages
          type: integer
        size:
          description: Items per page
          title: Size
          type: integer
        total:
          description: Total number of items
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - size
      - pages
      - has_next
      - has_prev
      title: PaginatedResponse
      type: object
...
    TicketAICreateRequest:
      description: Schema for AI-powered ticket creation
      properties:
        conversation_context:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          description: Conversation history
          title: Conversation Context
        integration:
          anyOf:
          - type: string
          - type: 'null'
          description: Integration name for external ticket creation
          title: Integration
        uploaded_files:
          anyOf:
          - items:
              format: uuid
              type: string
            type: array
          - type: 'null'
          description: Uploaded file IDs
          title: Uploaded Files
        user_input:
          description: User's natural language input
          title: User Input
          type: string
        user_preferences:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: User preferences
          title: User Preferences
      required:
      - user_input
      title: TicketAICreateRequest
      type: object
    TicketAICreateResponse:
      description: Response from AI ticket creation
      properties:
        ai_analysis:
          $ref: '#/components/schemas/TicketAIAnalysis'
          description: AI analysis results
        confidence_score:
          description: Overall confidence score
          title: Confidence Score
          type: number
        file_analysis_summary:
          anyOf:
          - type: string
          - type: 'null'
          description: File analysis summary
          title: File Analysis Summary
        suggested_actions:
          description: Suggested next actions
          items:
            type: string
          title: Suggested Actions
          type: array
        ticket:
          $ref: '#/components/schemas/TicketDetailResponse'
          description: Created ticket
      required:
      - ticket
      - ai_analysis
      - confidence_score
      - suggested_actions
      title: TicketAICreateResponse
      type: object
...
    TicketCategorySchema:
      description: Ticket category enum schema
      enum:
      - technical
      - billing
      - feature_request
      - bug
      - user_access
      - general
      - integration
      - performance
      - security
      title: TicketCategorySchema
      type: string
    TicketCreateRequest:
      description: Schema for creating a new ticket
      properties:
        assigned_to_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Assign to specific user
          title: Assigned To Id
        category:
          $ref: '#/components/schemas/TicketCategorySchema'
          default: general
          description: Issue category
        create_externally:
          default: true
          description: Create ticket in external system when integration specified
          title: Create Externally
          type: boolean
        custom_fields:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Custom field values
          title: Custom Fields
        department:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          description: Target department
          title: Department
        description:
          description: Detailed ticket description
          title: Description
          type: string
        file_ids:
          anyOf:
          - items:
              format: uuid
              type: string
            type: array
          - type: 'null'
          description: Attached file IDs
          title: File Ids
        integration:
          anyOf:
          - type: string
          - type: 'null'
          description: Integration name for external ticket creation
          title: Integration
        priority:
          $ref: '#/components/schemas/TicketPrioritySchema'
          default: medium
          description: Priority level
        title:
          description: Ticket title/subject
          maxLength: 500
          title: Title
          type: string
        urgency:
          $ref: '#/components/schemas/TicketPrioritySchema'
          default: medium
          description: Urgency level
      required:
      - title
      - description
      title: TicketCreateRequest
      type: object
...
    TicketPrioritySchema:
      description: Ticket priority enum schema
      enum:
      - low
      - medium
      - high
      - critical
      title: TicketPrioritySchema
      type: string
...
    TicketStatusSchema:
      description: Ticket status enum schema
      enum:
      - new
      - open
      - in_progress
      - pending
      - resolved
      - closed
      - cancelled
      title: TicketStatusSchema
      type: string
...
    TicketUpdateRequest:
      description: Schema for updating a ticket
      properties:
        assigned_to_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Assigned user ID
          title: Assigned To Id
        category:
          anyOf:
          - $ref: '#/components/schemas/TicketCategorySchema'
          - type: 'null'
          description: Issue category
        custom_fields:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Custom field values
          title: Custom Fields
        department:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          description: Department
          title: Department
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Ticket description
          title: Description
        integration:
          anyOf:
          - type: string
          - type: 'null'
          description: Integration for external ticket routing
          title: Integration
        internal_notes:
          anyOf:
          - type: string
          - type: 'null'
          description: Internal notes
          title: Internal Notes
        priority:
          anyOf:
          - $ref: '#/components/schemas/TicketPrioritySchema'
          - type: 'null'
          description: Priority level
        resolution_summary:
          anyOf:
          - type: string
          - type: 'null'
          description: Resolution summary
          title: Resolution Summary
        status:
          anyOf:
          - $ref: '#/components/schemas/TicketStatusSchema'
          - type: 'null'
          description: Ticket status
        title:
          anyOf:
          - maxLength: 500
            type: string
          - type: 'null'
          description: Ticket title
          title: Title
        urgency:
          anyOf:
          - $ref: '#/components/schemas/TicketPrioritySchema'
          - type: 'null'
          description: Urgency level
      title: TicketUpdateRequest
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: "\n        # AI Ticket Creator Backend API\n        \n        Comprehensive\
    \ API for AI-powered ticket creation and management with Chrome extension support.\n\
    \        \n        ## Features\n        \n        - **AI-Powered Ticket Creation**:\
    \ Smart ticket creation with context analysis\n        - **File Processing**:\
    \ Upload and process multiple file types with AI transcription/OCR\n        -\
    \ **Real-time Notifications**: WebSocket-based real-time updates\n        - **Third-party\
    \ Integrations**: Salesforce, Jira, ServiceNow, Zendesk, GitHub, Slack, Teams,\
    \ Zoom\n        - **Chrome Extension Support**: Optimized for browser extension\
    \ workflows\n        - **Multi-Tenant Organization Support**: Complete organization\
    \ isolation for enterprise deployments\n        \n        ## Quick Start Guide\n\
    \        \n        ### 1. Authentication\n        Register a new user with organization:\n\
    \        ```bash\n        curl -X POST \"http://localhost:8000/api/v1/auth/register\"\
    \ \\\n          -H \"Content-Type: application/json\" \\\n          -d '{\"email\"\
    : \"user@company.com\", \"full_name\": \"John Doe\", \"password\": \"SecurePass123\"\
    , \"organization_name\": \"Acme Corporation\"}'\n        ```\n        \n     \
    \   ### 2. Add JIRA Integration\n        ```bash\n        curl -X POST \"http://localhost:8000/api/v1/integrations\"\
...
    \ and meeting management\n        "
...
openapi: 3.1.0
paths:
...
  /api/v1/tickets/:
    get:
      description: List tickets with search, filtering, and pagination.
      operationId: list_tickets_api_v1_tickets__get
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Size
          type: integer
      - in: query
        name: q
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - in: query
        name: department
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
      - in: query
        name: created_by_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Created By Id
      - in: query
        name: assigned_to_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Assigned To Id
      - in: query
        name: integration
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Integration
      - in: query
        name: has_attachments
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Attachments
      - in: query
        name: is_overdue
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Overdue
      - in: query
        name: escalation_level
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Escalation Level
      - in: query
        name: satisfaction_rating
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Satisfaction Rating
      - in: query
        name: created_after
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created After
      - in: query
        name: created_before
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created Before
      - in: query
        name: updated_after
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated After
      - in: query
        name: updated_before
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated Before
      - in: query
        name: resolved_after
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Resolved After
      - in: query
        name: resolved_before
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Resolved Before
      - in: query
        name: source_channel
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Channel
      - in: query
        name: sort_by
        required: false
        schema:
          default: created_at
          pattern: ^(created_at|updated_at|title|priority|status|last_activity_at|resolution_time_minutes)$
          title: Sort By
          type: string
      - in: query
        name: sort_order
        required: false
        schema:
          default: desc
          pattern: ^(asc|desc)$
          title: Sort Order
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_list_tickets_api_v1_tickets__get'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Tickets
      tags:
      - Tickets
    post:
      description: Create a new support ticket.
      operationId: create_ticket_api_v1_tickets__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create Ticket
      tags:
      - Tickets
  /api/v1/tickets/ai-create:
    post:
      description: Create a support ticket using AI analysis of user input.
      operationId: create_ticket_with_ai_api_v1_tickets_ai_create_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketAICreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketAICreateResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create Ticket With Ai
      tags:
      - Tickets
...
  /api/v1/tickets/{ticket_id}:
    delete:
      description: Soft delete a ticket.
      operationId: delete_ticket_api_v1_tickets__ticket_id__delete
      parameters:
      - in: path
        name: ticket_id
        required: true
        schema:
          format: uuid
          title: Ticket Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Ticket
      tags:
      - Tickets
    get:
      description: Get a specific ticket by ID.
      operationId: get_ticket_api_v1_tickets__ticket_id__get
      parameters:
      - in: path
        name: ticket_id
        required: true
        schema:
          format: uuid
          title: Ticket Id
          type: string
      - description: Include AI analysis data
        in: query
        name: include_ai_data
        required: false
        schema:
          default: true
          description: Include AI analysis data
          title: Include Ai Data
          type: boolean
      - description: Include internal notes
        in: query
        name: include_internal
        required: false
        schema:
          default: false
          description: Include internal notes
          title: Include Internal
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Ticket
      tags:
      - Tickets
    put:
      description: Update a ticket.
      operationId: update_ticket_api_v1_tickets__ticket_id__put
      parameters:
      - in: path
        name: ticket_id
        required: true
        schema:
          format: uuid
          title: Ticket Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Ticket
      tags:
      - Tickets
...